pyproject.toml
[tool.poetry]
name = "repo-summarizer"
version = "0.1.0"
description = ""
authors = ["Martin Hinrichs <mail@martinhinrichs.net>"]
readme = "README.md"
packages = [{include = "repo_summarizer"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
langchain = "^0.0.144"
openai = "^0.27.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
